Address the latest review comments by: 1. delete constants.go and move NoGID to fileutil 2. Re-include "volume mount with rotation but skipped" in nodeserver_test.go 3. Added comment about why 64 bit word was chosen for FsGroup in filesystem.go 4. Change the GID and FSGroup to generic int to allow for 64bit int if and when applicable    as provisioned at https://github.com/kubernetes/kubernetes/blob/b910026535af2d8a64d45efefeb8d9efb75a4817/pkg/volume/csi/csi_client.go#L64 5. Fixed enable_secret_rotation in e2e-provider.bats to correctly return the name of the pod.

+++ pkg/secrets-store/nodeserver.go
@@ -350,7 +350,7 @@ func (ns *nodeServer) NodeUnstageVolume(ctx context.Context, req *csi.NodeUnstag
 	return &csi.NodeUnstageVolumeResponse{}, nil
 }
 
-func (ns *nodeServer) mountSecretsStoreObjectContent(ctx context.Context, providerName, attributes, secrets, targetPath, permission, podName string, gid int64) (map[string]string, string, error) {
+func (ns *nodeServer) mountSecretsStoreObjectContent(ctx context.Context, providerName, attributes, secrets, targetPath, permission, podName string, gid int) (map[string]string, string, error) {
 	if len(attributes) == 0 {
 		return nil, "", errors.New("missing attributes")
 	}
